Seperate the data verification functions into their own file
